requirements.txt
cvxpy==1.2.0
numpy
PyYAML==5.4.1
qpth==0.0.15
rich==12.2.0
tensorboard-logger==0.1.0
tensorboardX==2.5
torch==1.10.1+cu111
tqdm
